+++ src/dcrypt/keytool.c
@@ -99,7 +99,7 @@ void read_keyfile(struct keyfile* kf, char* name)
 
 	if((ret = sys_read(fd, kf->buf, st.size)) < 0)
 		fail("read", name, ret);
-	if(ret < st.size)
+	if((ulong)ret < st.size)
 		fail("incomplete read", NULL, 0);
 	if(ret < 16 || ret % 16)
 		fail("invalid keyfile", name, 0);
